requirements.txt
django
pillow
gunicorn==21.2.0
django-filter
django-jazzmin==2.6.0